Factor API URL detection and support URLs with paths
