+++ src/renderer/components/Login/LoginPage.tsx
@@ -9,7 +9,7 @@ import {
   ModalDialog,
   Typography,
 } from '@mui/joy';
-import { getPath } from 'renderer/lib/api-client/urls';
+import { deriveApiBaseUrl, getPath } from 'renderer/lib/api-client/urls';
 import { Routes, Route, useNavigate } from 'react-router-dom';
 import { useAuth } from '../../lib/authContext';
 import HexLogo from '../Shared/HexLogo';
@@ -284,38 +284,7 @@ export default function LoginPage() {
     if (token) {
       const verifyEmail = async () => {
         try {
-          // Normalize TL_API_URL - ensure it's not "default" or empty, and mirror App.tsx behavior:
-          // - For localhost or port 1212, assume API is on port 8338
-          // - For non-localhost, assume API is served from the same origin as the frontend
-          const envUrl = process.env.TL_API_URL;
-          let apiUrl: string;
-
-          if (!envUrl || envUrl === 'default' || envUrl.trim() === '') {
-            const { protocol, hostname, port } = window.location;
-
-            if (hostname === 'localhost' || hostname === '127.0.0.1') {
-              apiUrl = `${protocol}//${hostname}:8338`;
-            } else if (port === '1212') {
-              apiUrl = `${protocol}//${hostname}:8338`;
-            } else {
-              const isDefaultHttpPort = port === '' || port === '80';
-              const isDefaultHttpsPort = port === '' || port === '443';
-              const isDefaultPort =
-                (protocol === 'http:' && isDefaultHttpPort) ||
-                (protocol === 'https:' && isDefaultHttpsPort);
-
-              if (isDefaultPort) {
-                apiUrl = `${protocol}//${hostname}`;
-              } else {
-                apiUrl = `${protocol}//${hostname}:${port}`;
-              }
-            }
-          } else {
-            apiUrl = envUrl.trim();
-          }
-
-          apiUrl = apiUrl.replace(/\/$/, '');
-          const url = `${apiUrl}/auth/verify`;
+          const url = `${deriveApiBaseUrl()}auth/verify`;
           const response = await fetch(url, {
             method: 'POST',
             headers: {
