+++ src/renderer/lib/api-client/urls.ts
@@ -1,5 +1,50 @@
 import endpointsJson from './allEndpoints.json';
 
+// Resolve the API base URL. Always returns a string ending with '/'.
+//
+// Order of resolution:
+//   1. process.env.TL_API_URL (build-time, baked in by webpack). May be a full
+//      URL ('https://api.example.com/') or a path-only prefix ('/lab/').
+//   2. window.location-derived default — same origin as the frontend, with
+//      the SPA mount sub-path preserved (e.g. served at /tlab/ → /tlab/),
+//      so reverse-proxy path-prefix deployments work without TL_API_URL.
+//      Local dev (localhost / 127.0.0.1 / port 1212) falls back to :8338.
+//
+// Sub-path preservation relies on HashRouter — pathname stays at the original
+// SPA mount and only the hash changes during client-side navigation.
+export function deriveApiBaseUrl(): string {
+  const envUrl = process.env?.TL_API_URL;
+  if (envUrl && envUrl !== 'default' && envUrl.trim() !== '') {
+    let url = envUrl.trim();
+    if (!url.endsWith('/')) url += '/';
+    return url;
+  }
+
+  const { protocol, hostname, port, pathname } = window.location;
+
+  if (hostname === 'localhost' || hostname === '127.0.0.1') {
+    return `${protocol}//${hostname}:8338/`;
+  }
+  if (port === '1212') {
+    return `${protocol}//${hostname}:8338/`;
+  }
+
+  // Trim pathname back to the last '/' so /tlab/index.html and /tlab/ both
+  // resolve to /tlab/, and / stays /. HashRouter never modifies pathname,
+  // so this captures the SPA mount even after in-app navigation.
+  const lastSlash = pathname.lastIndexOf('/');
+  const basePath = lastSlash >= 0 ? pathname.slice(0, lastSlash + 1) : '/';
+
+  const isDefaultHttpPort = port === '' || port === '80';
+  const isDefaultHttpsPort = port === '' || port === '443';
+  const isDefaultPort =
+    (protocol === 'http:' && isDefaultHttpPort) ||
+    (protocol === 'https:' && isDefaultHttpsPort);
+
+  const host = isDefaultPort ? hostname : `${hostname}:${port}`;
+  return `${protocol}//${host}${basePath}`;
+}
+
 export function API_URL() {
   const raw = (window as any).TransformerLab?.API_URL || null;
   if (!raw) return null;
