+++ src/renderer/lib/authContext.ts
@@ -8,7 +8,7 @@ import React, {
 } from 'react';
 import { API_URL } from './transformerlab-api-sdk';
 import useSWR from 'swr';
-import { getAPIFullPath, getPath } from './api-client/urls';
+import { deriveApiBaseUrl, getAPIFullPath, getPath } from './api-client/urls';
 import {
   identifyUser,
   resetUser,
@@ -200,21 +200,15 @@ export async function fetchWithAuth(url: string, options: RequestInit = {}) {
   // Handle cases where url might be partial or full
   // Ideally fetchWithAuth is passed a relative path, but we handle both.
   let fullUrl: string;
-  if (url.startsWith('http')) {
+  if (url.startsWith('http') || url.startsWith('//') || url.startsWith('/')) {
+    // Already an absolute URL (http://...) or absolute path (/lab/...).
+    // getAPIFullPath returns the latter when API_URL is configured as a path
+    // prefix (e.g. when the app is hosted behind a reverse proxy at /lab/).
     fullUrl = url;
   } else {
-    const baseUrl = API_URL();
-    if (baseUrl === null) {
-      // Default to same host as frontend with API port if API_URL is not set
-      // Ensure URL doesn't start with / (baseUrl already has trailing slash)
-      const cleanPath = url.startsWith('/') ? url.slice(1) : url;
-      const protocol = window.location.protocol;
-      const hostname = window.location.hostname;
-      fullUrl = `${protocol}//${hostname}:8338/${cleanPath}`;
-    } else {
-      // baseUrl already has trailing slash from API_URL()
-      fullUrl = `${baseUrl}${url}`;
-    }
+    const baseUrl = API_URL() ?? deriveApiBaseUrl();
+    // baseUrl always has a trailing slash
+    fullUrl = `${baseUrl}${url}`;
   }
 
   const method = (options.method ?? 'GET').toUpperCase();
